+++ tests/test_dataset_routes.py
@@ -444,3 +444,21 @@ def test_dataset_manifest_route(tmp_app_with_data):  # NOQA
         content_type="application/json"
     )
     assert r.status_code == 400
+
+
+def test_dataset_manifest_route(tmp_app_with_data):  # NOQA
+
+    headers = dict(Authorization="Bearer " + grumpy_token)
+    query = {"uri": "s3://snow-white/af6727bf-29c7-43dd-b42f-a5d7ede28337"}
+    r = tmp_app_with_data.post(
+        "/dataset/readme",
+        headers=headers,
+        data=json.dumps(query),
+        content_type="application/json"
+    )
+    assert r.status_code == 200
+
+    expected_readme = {"descripton": "apples from queen"}
+    actual_readme = json.loads(r.data.decode("utf-8"))
+
+    assert expected_readme == actual_readme